# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Aman APIs
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Ingest Service
  - url: http://localhost:8081
    description: UI API
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /v1/decisions:
    get:
      summary: Decisions list (filterable)
      parameters:
        - in: query
          name: verdict
          schema:
            type: string
        - in: query
          name: rule_id
          schema:
            type: string
        - in: query
          name: thread_id
          schema:
            type: string
        - in: query
          name: confidence_min
          schema:
            type: number
        - in: query
          name: confidence_max
          schema:
            type: number
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: Decisions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionsEnvelope"
...
  /v1/feedback:
    post:
      summary: Submit analyst feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AckEnvelope"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Meta:
      type: object
      properties:
        request_id:
          type: string
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    DecisionItem:
      type: object
      properties:
        rule_id:
          type: string
        name:
          type: string
        verdict:
          type: string
        confidence:
          type: number
        reason_code:
          type: string
        decision_label:
          type: string
        thread_id:
          type: string
        missing_evidence:
          type: array
          items:
            type: string
        updated:
          type: string
        evidence:
          type: array
          items:
            type: string
        explanation:
          type: string
    DecisionsEnvelope:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DecisionItem"
        meta:
          $ref: "#/components/schemas/Meta"
...
    AckEnvelope:
      type: object
      properties:
        data:
          type: object
        meta:
          $ref: "#/components/schemas/Meta"